feat(Checkpoint): add checkpoint
