+++ Assets/Scripts/Interactable/Checkpoint.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c15c6281d94a7947a30b47ef4cd102f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
