+++ Assets/Scripts/Interactable/Checkpoint/Checkpoint.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Checkpoint : Interactable
+{
+    public override void Interact()
+    {
+        // TODO: Implement interaction with portal
+        Debug.Log("Trigger Checkpoint interaction");
+    }
+}
