+++ Assets/Scripts/Interactable/Checkpoint/Checkpoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48d37faf585ea794bb02ee2a95fdbad0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
