+++ Assets/Scripts/Interactable/Portal/PortalController.cs
@@ -3,19 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class PortalController : Interactable
-{
-    public ControllableComponent controller;
-
-    void OnTriggerEnter2D(Collider2D other)
-    {
-        controller.SignalIsOnInteractable(this);
-    }
-
-    void OnTriggerExit2D(Collider2D other)
-    {
-        controller.SignalIsOnInteractable(null);
-    }
-    
+{    
     public override void Interact()
     {
         // TODO: Implement interaction with portal
